adiciona base para tradução na homepage
